+++ react-shopping-cart/src/components/@shared/index.js
@@ -1,0 +1,8 @@
+import Error from './Error/Error.component';
+import HeaderContainer from './HeaderContainer/HeaderContainer.component';
+import HeaderLink from './HeaderLink/HeaderLink.component';
+import Image from './Image/Image.component';
+import PageContainer from './PageContainer/PageContainer.component';
+import Text from './Text/Text.component';
+
+export { Error, HeaderContainer, HeaderLink, Image, PageContainer, Text };
