+++ react-shopping-cart/src/pages/index.js
@@ -1,0 +1,4 @@
+import ProductList from './ProductList/ProductList.page';
+import NotFound from './NotFound/NotFound.page';
+
+export { ProductList, NotFound };
