+++ react-shopping-cart/src/styles/index.js
@@ -1,0 +1,4 @@
+import GlobalStyle from './GlobalStyle';
+import theme from './theme';
+
+export { GlobalStyle, theme };
